Уникальные id комментариев и вынос вспомогательных функций в другие модули.
