+++ index.html
@@ -233,5 +233,5 @@
   </template>
 
 </body>
-<script src="js/main.js"></script>
+<script type="module" src="js/main.js"></script>
 </html>
