+++ js/util.js
@@ -1,0 +1,5 @@
+const getRandom = (firstNumber, lastNumber) => Math.round(Math.random() * (lastNumber - firstNumber) + firstNumber);
+const checkWordsCount = (wordsLine, maxSymbol) => wordsLine.length <= maxSymbol;
+
+
+export {getRandom,checkWordsCount};
